requirements.txt
streamlit
pyrebase4
firebase-admin